services/articles/requirements.txt
Werkzeug==0.14.1 # this version can keep OSError: [Errno 8] Exec format error occurring
autopep8==1.4.4
coverage==4.5.2
flake8==3.6.0
Flask==1.0.2
Flask-Cors==3.0.7
Flask-DebugToolbar==0.10.1
Flask-Migrate==2.4.0
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.7.1
Jinja2==2.10
PyJWT==1.7.1
pylint==2.3.1
psycopg2-binary==2.7.6.1
gunicorn==19.9.0